fix(complete)!: Rename shells to command
