+++ clap_complete/src/dynamic/mod.rs
@@ -1,7 +1,5 @@
 //! Complete commands within shells
 //!
-//! For quick-start, see [`CompleteCommand`]
-//!
 //! To customize completions, see
 //! - [`ValueHint`][crate::ValueHint]
 //! - [`ValueEnum`][clap::ValueEnum]
@@ -11,15 +9,15 @@ mod candidate;
 mod complete;
 mod custom;
 
-pub mod shells;
+#[cfg(feature = "unstable-command")]
+pub mod command;
 
 pub use candidate::CompletionCandidate;
 pub use complete::complete;
 pub use custom::ArgValueCompleter;
 pub use custom::CustomCompleter;
 
-// These live in `shells` because they are tightly coupled with the `ShellCompleter`s
 #[cfg(feature = "unstable-command")]
-pub use shells::CompleteArgs;
+pub use command::CompleteArgs;
 #[cfg(feature = "unstable-command")]
-pub use shells::CompleteCommand;
+pub use command::CompleteCommand;
