+++ clap_complete/tests/testsuite/bash.rs
@@ -114,14 +114,14 @@ fn value_terminator() {
 #[cfg(feature = "unstable-command")]
 #[test]
 fn register_minimal() {
-    use clap_complete::dynamic::shells::CommandCompleter as _;
+    use clap_complete::dynamic::command::CommandCompleter as _;
 
     let name = "my-app";
     let bin = name;
     let completer = name;
 
     let mut buf = Vec::new();
-    clap_complete::dynamic::shells::Bash
+    clap_complete::dynamic::command::Bash
         .write_registration(name, bin, completer, &mut buf)
         .unwrap();
     snapbox::Assert::new()
